Delete duplicate code transparency-request
